Snapshot playlist contents
Loop through all playlists and take a snapshot of their contents in order of when the tracks were added. Each playlist gets its own file in the snapshots repo, formatted in the same way as the saved tracks file at the moment. Eventually we'll probably start snapshotting who added a track to a playlist, for collaborative ones.