Make first snapshot commit message better
Additions/deletions doesn't make sense for the first commit, so just have the message describe repo state

This is not 100% accurate (had some off-by-one count errors while testing) but honestly it's close enough

The better way would probably be to actually open and read the files written, rather than looking at commit data. But this is fine for now.